Draw background image

+++ backend/showcase/src/consts.rs
@@ -1,1 +1,1 @@
-pub const PIXELS_PER_TILE: u32 = 20;
+pub const PIXELS_PER_TILE: usize = 20;
